fix(simulator): share implementation with browser demo
